# packages/cdk/openapi.json
{"openapi":"3.0.0","info":{"title":"Akello","description":"API's for all of Akello's core services","version":"1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/registry/create":{"post":{"tags":["Registry"],"summary":"Create Registry","operationId":"create_registry_v1_registry_create_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/registry/{registry_id}":{"get":{"tags":["Registry"],"summary":"Get Registry","operationId":"get_registry_v1_registry__registry_id__get","parameters":[{"name":"registry_id","in":"path","required":true,"schema":{"type":"string","title":"Registry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/registry/{registry_id}/team-members":{"get":{"tags":["Registry"],"summary":"Get Registry Team Members","operationId":"get_registry_team_members_v1_registry__registry_id__team_members_get","parameters":[{"name":"registry_id","in":"path","required":true,"schema":{"type":"string","title":"Registry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/registry/{registry_id}/patients":{"get":{"tags":["Registry"],"summary":"Get Registry Patients","operationId":"get_registry_patients_v1_registry__registry_id__patients_get","parameters":[{"name":"registry_id","in":"path","required":true,"schema":{"type":"string","title":"Registry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/registry/{registry_id}/refer-patient":{"post":{"tags":["Registry"],"summary":"Refer Patient","operationId":"refer_patient_v1_registry__registry_id__refer_patient_post","parameters":[{"name":"registry_id","in":"path","required":true,"schema":{"type":"string","title":"Registry Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientRegistry"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/registry/{registry_id}/record-session":{"post":{"tags":["Registry"],"summary":"Record Session","operationId":"record_session_v1_registry__registry_id__record_session_post","parameters":[{"name":"registry_id","in":"path","required":true,"schema":{"type":"string","title":"Registry Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TreatmentLog"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/registry/{registry_id}/patient-attribute":{"post":{"tags":["Registry"],"summary":"Set Patient Attribute","operationId":"set_patient_attribute_v1_registry__registry_id__patient_attribute_post","parameters":[{"name":"registry_id","in":"path","required":true,"schema":{"type":"string","title":"Registry Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Data"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/user":{"get":{"tags":["User"],"summary":"Get User","operationId":"get_user_v1_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/user/registries":{"get":{"tags":["User"],"summary":"Get User Registries","operationId":"get_user_registries_v1_user_registries_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/financial-model":{"get":{"tags":["Financial Model"],"summary":"Get Models","operationId":"get_models_v1_financial_model_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/FinancialModel"},"type":"array","title":"Response Get Models V1 Financial Model Get"}}}}}},"put":{"tags":["Financial Model"],"summary":"Put Model","description":"financial_model = load_client_model(model, auth.cognito_id)\nquery = FinancialModelQuery()\nquery.create_model(financial_model)","operationId":"put_model_v1_financial_model_put","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Model"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Financial Model"],"summary":"Create Model","description":"financial_model = load_db_model(model, auth.cognito_id)\nquery = FinancialModelQuery()\nquery.create_model(financial_model)","operationId":"create_model_v1_financial_model_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Model"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/financial-model/{model_name}":{"get":{"tags":["Financial Model"],"summary":"Get Model","operationId":"get_model_v1_financial_model__model_name__get","parameters":[{"name":"model_name","in":"path","required":true,"schema":{"title":"Model Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FinancialModel"},"title":"Response Get Model V1 Financial Model  Model Name  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/reports/{registry_id}/billing":{"get":{"tags":["Reports"],"summary":"Billing Report Monthly","operationId":"billing_report_monthly_v1_reports__registry_id__billing_get","parameters":[{"name":"registry_id","in":"path","required":true,"schema":{"type":"string","title":"Registry Id"}},{"name":"from_date","in":"query","required":true,"schema":{"type":"integer","title":"From Date"}},{"name":"to_date","in":"query","required":true,"schema":{"type":"integer","title":"To Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/reports/{registry_id}/dashboard-stats":{"get":{"tags":["Reports"],"summary":"Get Registry Stats","operationId":"get_registry_stats_v1_reports__registry_id__dashboard_stats_get","parameters":[{"name":"registry_id","in":"path","required":true,"schema":{"type":"string","title":"Registry Id"}},{"name":"from_date","in":"query","required":true,"schema":{"type":"integer","title":"From Date"}},{"name":"to_date","in":"query","required":true,"schema":{"type":"integer","title":"To Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuditLog":{"properties":{"object_name":{"type":"string","title":"Object Name"},"action":{"type":"string","title":"Action"},"user_id":{"type":"string","title":"User Id"},"created_date":{"type":"number","title":"Created Date"}},"type":"object","required":["object_name","action","user_id","created_date"],"title":"AuditLog"},"ContactTypes":{"type":"string","enum":["Initial Assessment","Follow Up","Psychiatric Consultation","Relapse Prevention Plan"],"title":"ContactTypes"},"EventLog":{"properties":{"id":{"type":"string","title":"Id"},"system":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"System"},"data":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Data"},"created_date":{"type":"number","title":"Created Date"},"modified_date":{"type":"number","title":"Modified Date"}},"type":"object","required":["id","created_date","modified_date"],"title":"EventLog"},"FinancialModel":{"properties":{"schema_version":{"type":"string","title":"Schema Version"},"user_id":{"type":"string","title":"User Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"address_state":{"type":"string","title":"Address State"},"full_time_pcps":{"type":"integer","title":"Full Time Pcps"},"pct_pcps_participation":{"type":"number","title":"Pct Pcps Participation"},"monthly_pcp_referrals":{"type":"integer","title":"Monthly Pcp Referrals"},"pct_patients_accepting_cocm":{"type":"number","title":"Pct Patients Accepting Cocm"},"pct_patients_graduating":{"type":"number","title":"Pct Patients Graduating"},"payer_distribution_medicare":{"type":"number","title":"Payer Distribution Medicare"},"payer_distribution_medicare_cocm_expected":{"type":"number","title":"Payer Distribution Medicare Cocm Expected"},"payer_distribution_medicaid":{"type":"number","title":"Payer Distribution Medicaid"},"payer_distribution_medicaid_cocm_expected":{"type":"number","title":"Payer Distribution Medicaid Cocm Expected"},"payer_distribution_commercial_cocm":{"type":"number","title":"Payer Distribution Commercial Cocm"},"payer_distribution_commercial_cocm_cocm_expected":{"type":"number","title":"Payer Distribution Commercial Cocm Cocm Expected"},"payer_distribution_commercial_no_cocm":{"type":"number","title":"Payer Distribution Commercial No Cocm"},"caseload_complexity_pct_patients_suicidality":{"type":"number","title":"Caseload Complexity Pct Patients Suicidality","default":0},"caseload_complexity_pct_patients_prior_ed":{"type":"number","title":"Caseload Complexity Pct Patients Prior Ed","default":0},"caseload_complexity_pct_patients_substance_abuse":{"type":"number","title":"Caseload Complexity Pct Patients Substance Abuse","default":0},"caseload_complexity_pct_patients_above_moderate":{"type":"number","title":"Caseload Complexity Pct Patients Above Moderate","default":0},"care_manager_salary":{"type":"number","title":"Care Manager Salary"},"care_manager_benefits":{"type":"number","title":"Care Manager Benefits"},"psychiatrist_salary":{"type":"number","title":"Psychiatrist Salary"},"psychiatrist_benefits":{"type":"number","title":"Psychiatrist Benefits"},"service_unit_direct_bhm__warm_connection_over_16":{"type":"number","title":"Service Unit Direct Bhm  Warm Connection Over 16"},"service_unit_direct_bhm__initial_assessment_visit":{"type":"number","title":"Service Unit Direct Bhm  Initial Assessment Visit"},"service_unit_direct_bhm__follow_up_visit":{"type":"number","title":"Service Unit Direct Bhm  Follow Up Visit"},"service_unit_direct_bhm__group_treatment":{"type":"number","title":"Service Unit Direct Bhm  Group Treatment"},"service_unit_other_bhm__warm_connection_under1_16":{"type":"number","title":"Service Unit Other Bhm  Warm Connection Under1 16"},"service_unit_other_bhm__outreach_attempts":{"type":"number","title":"Service Unit Other Bhm  Outreach Attempts"},"service_unit_other_bhm__telephone_visit":{"type":"number","title":"Service Unit Other Bhm  Telephone Visit"},"service_unit_other_bhm__caseload_patient_and_psych_consult":{"type":"number","title":"Service Unit Other Bhm  Caseload Patient And Psych Consult"},"service_unit_other_bhm__team_communication":{"type":"number","title":"Service Unit Other Bhm  Team Communication"},"service_unit_other_bhm__registry_management":{"type":"number","title":"Service Unit Other Bhm  Registry Management"},"service_unit_admin_bhm__charting":{"type":"number","title":"Service Unit Admin Bhm  Charting"},"service_unit_admin_bhm__other":{"type":"number","title":"Service Unit Admin Bhm  Other"},"service_unit_direct_cp_treatment_visit":{"type":"number","title":"Service Unit Direct Cp Treatment Visit"},"service_unit_direct_cp_follow_up":{"type":"number","title":"Service Unit Direct Cp Follow Up"},"service_unit_other_cp_registry_review":{"type":"number","title":"Service Unit Other Cp Registry Review"},"service_unit_other_cp_direct_pcp_communication":{"type":"number","title":"Service Unit Other Cp Direct Pcp Communication"},"service_unit_other_cp_caseload_review":{"type":"number","title":"Service Unit Other Cp Caseload Review"},"service_unit_other_cp_charting":{"type":"number","title":"Service Unit Other Cp Charting"},"service_unit_other_cp_other":{"type":"number","title":"Service Unit Other Cp Other"},"billing_rate_not_seen__patients":{"type":"number","title":"Billing Rate Not Seen  Patients"},"billing_rate_not_seen__avg_amount":{"type":"number","title":"Billing Rate Not Seen  Avg Amount"},"billing_rate_G2214__patients":{"type":"number","title":"Billing Rate G2214  Patients"},"billing_rate_G2214__avg_amount":{"type":"number","title":"Billing Rate G2214  Avg Amount"},"billing_rate_99492__patients":{"type":"number","title":"Billing Rate 99492  Patients"},"billing_rate_99492__avg_amount":{"type":"number","title":"Billing Rate 99492  Avg Amount"},"billing_rate_99492_99494__patients":{"type":"number","title":"Billing Rate 99492 99494  Patients"},"billing_rate_99492_99494__avg_amount":{"type":"number","title":"Billing Rate 99492 99494  Avg Amount"},"billing_rate_99492_99494x2__patients":{"type":"number","title":"Billing Rate 99492 99494X2  Patients"},"billing_rate_99492_99494x2__avg_amount":{"type":"number","title":"Billing Rate 99492 99494X2  Avg Amount"},"billing_rate_99493__patients":{"type":"number","title":"Billing Rate 99493  Patients"},"billing_rate_99493__avg_amount":{"type":"number","title":"Billing Rate 99493  Avg Amount"},"billing_rate_99493_99494__patients":{"type":"number","title":"Billing Rate 99493 99494  Patients"},"billing_rate_99493_99494__avg_amount":{"type":"number","title":"Billing Rate 99493 99494  Avg Amount"},"billing_rate_99493_99494x2__patients":{"type":"number","title":"Billing Rate 99493 99494X2  Patients"},"billing_rate_99493_99494x2__avg_amount":{"type":"number","title":"Billing Rate 99493 99494X2  Avg Amount"},"modified_date":{"type":"number","title":"Modified Date"},"created_date":{"type":"number","title":"Created Date"}},"type":"object","required":["schema_version","user_id","name","description","address_state","full_time_pcps","pct_pcps_participation","monthly_pcp_referrals","pct_patients_accepting_cocm","pct_patients_graduating","payer_distribution_medicare","payer_distribution_medicare_cocm_expected","payer_distribution_medicaid","payer_distribution_medicaid_cocm_expected","payer_distribution_commercial_cocm","payer_distribution_commercial_cocm_cocm_expected","payer_distribution_commercial_no_cocm","care_manager_salary","care_manager_benefits","psychiatrist_salary","psychiatrist_benefits","service_unit_direct_bhm__warm_connection_over_16","service_unit_direct_bhm__initial_assessment_visit","service_unit_direct_bhm__follow_up_visit","service_unit_direct_bhm__group_treatment","service_unit_other_bhm__warm_connection_under1_16","service_unit_other_bhm__outreach_attempts","service_unit_other_bhm__telephone_visit","service_unit_other_bhm__caseload_patient_and_psych_consult","service_unit_other_bhm__team_communication","service_unit_other_bhm__registry_management","service_unit_admin_bhm__charting","service_unit_admin_bhm__other","service_unit_direct_cp_treatment_visit","service_unit_direct_cp_follow_up","service_unit_other_cp_registry_review","service_unit_other_cp_direct_pcp_communication","service_unit_other_cp_caseload_review","service_unit_other_cp_charting","service_unit_other_cp_other","billing_rate_not_seen__patients","billing_rate_not_seen__avg_amount","billing_rate_G2214__patients","billing_rate_G2214__avg_amount","billing_rate_99492__patients","billing_rate_99492__avg_amount","billing_rate_99492_99494__patients","billing_rate_99492_99494__avg_amount","billing_rate_99492_99494x2__patients","billing_rate_99492_99494x2__avg_amount","billing_rate_99493__patients","billing_rate_99493__avg_amount","billing_rate_99493_99494__patients","billing_rate_99493_99494__avg_amount","billing_rate_99493_99494x2__patients","billing_rate_99493_99494x2__avg_amount","modified_date","created_date"],"title":"FinancialModel","examples":[{"name":"Foo"}]},"FlagTypes":{"type":"string","enum":["Needs Discussion","Review with Psychiatrist","Safety Risk"],"title":"FlagTypes"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PatientRegistry":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"patient_flag":{"anyOf":[{"$ref":"#/components/schemas/FlagTypes"},{"type":"null"}]},"patient_mrn":{"type":"string","title":"Patient Mrn"},"date_created":{"type":"number","title":"Date Created","default":1707989895.857551},"date_graduated":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Date Graduated"},"date_modified":{"type":"number","title":"Date Modified","default":1707989895.857558},"payer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Payer"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"phone_number":{"type":"string","title":"Phone Number"},"email":{"type":"string","title":"Email"},"date_of_birth":{"type":"string","title":"Date Of Birth"},"treatment_logs":{"items":{"$ref":"#/components/schemas/TreatmentLog"},"type":"array","title":"Treatment Logs","default":[]},"event_logs":{"items":{"$ref":"#/components/schemas/EventLog"},"type":"array","title":"Event Logs","default":[]},"audit_logs":{"items":{"$ref":"#/components/schemas/AuditLog"},"type":"array","title":"Audit Logs","default":[]},"flags":{"items":{"type":"object"},"type":"array","title":"Flags","default":[]},"status":{"allOf":[{"$ref":"#/components/schemas/PatientStatysTypes"}],"default":"Enrolled"},"initial_assessment":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Initial Assessment"},"last_follow_up":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Last Follow Up"},"last_psychiatric_consult":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Last Psychiatric Consult"},"relapse_prevention_plan":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Relapse Prevention Plan"},"total_sessions":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Sessions","default":0},"weeks_since_initial_assessment":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Weeks Since Initial Assessment","default":0},"minutes_this_month":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Minutes This Month","default":0},"schema_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Schema Version"}},"type":"object","required":["patient_mrn","first_name","last_name","phone_number","email","date_of_birth"],"title":"PatientRegistry","examples":[{"name":"Foo"}]},"PatientStatysTypes":{"type":"string","enum":["Enrolled","Treatment","Relapse Prevention Plan","Deactivated"],"title":"PatientStatysTypes"},"TreatmentLog":{"properties":{"patient_mrn":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Patient Mrn"},"provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider"},"no_show":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"No Show","default":false},"flag":{"anyOf":[{"$ref":"#/components/schemas/FlagTypes"},{"type":"null"}]},"weeks_in_treatment":{"type":"integer","title":"Weeks In Treatment"},"contact_type":{"$ref":"#/components/schemas/ContactTypes"},"visit_type":{"$ref":"#/components/schemas/VisitTypes"},"scores":{"items":{"$ref":"#/components/schemas/TreatmentLogScore"},"type":"array","title":"Scores","default":[]},"minutes":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Minutes"},"sms_reminder_sent_date":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Sms Reminder Sent Date"},"sms_conformation_received_date":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Sms Conformation Received Date"},"total_sms_reminders_sent":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Sms Reminders Sent","default":0},"email_reminder_sent_date":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Email Reminder Sent Date"},"email_conformation_received_date":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Email Conformation Received Date"},"total_email_reminders_sent":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Email Reminders Sent","default":0},"date":{"type":"number","title":"Date"}},"type":"object","required":["weeks_in_treatment","contact_type","visit_type","date"],"title":"TreatmentLog"},"TreatmentLogScore":{"properties":{"score_name":{"type":"string","title":"Score Name"},"score_value":{"type":"integer","title":"Score Value"}},"type":"object","required":["score_name","score_value"],"title":"TreatmentLogScore"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VisitTypes":{"type":"string","enum":["Clinic","Phone","In-person w/ Patient"],"title":"VisitTypes"}}}}